# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TechPath Jobs & Internships API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  student: "student",
  recruiter: "recruiter",
  admin: "admin",
} as const;

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
  role: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  recruiter: "recruiter",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
  profileComplete: boolean;
}

export interface AuthResponse {
  user: User;
  message?: string;
}

export interface Education {
  institution: string;
  degree: string;
  field: string;
  startYear: number;
  endYear?: number;
  gpa?: string;
  achievements?: string;
}

export interface Experience {
  company: string;
  role: string;
  location?: string;
  startDate: string;
  endDate?: string;
  current?: boolean;
  description?: string;
  techStack?: string[];
}

export interface Project {
  name: string;
  description: string;
  techStack?: string[];
  url?: string;
  githubUrl?: string;
  startDate?: string;
  endDate?: string;
}

export interface Certification {
  name: string;
  issuer: string;
  issueDate?: string;
  expiryDate?: string;
  credentialUrl?: string;
}

export interface Profile {
  id: number;
  userId: number;
  headline?: string;
  phone?: string;
  location?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  portfolioUrl?: string;
  summary?: string;
  skills: string[];
  education: Education[];
  experience: Experience[];
  projects: Project[];
  certifications: Certification[];
  languages?: string[];
  preferredRoles?: string[];
  preferredLocations?: string[];
  expectedSalary?: number;
  availableFrom?: string;
  openToRemote?: boolean;
  updatedAt?: string;
}

export interface ProfileInput {
  headline?: string;
  phone?: string;
  location?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  portfolioUrl?: string;
  summary?: string;
  skills: string[];
  education: Education[];
  experience: Experience[];
  projects: Project[];
  certifications: Certification[];
  languages?: string[];
  preferredRoles?: string[];
  preferredLocations?: string[];
  expectedSalary?: number;
  availableFrom?: string;
  openToRemote?: boolean;
}

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  internship: "internship",
  fulltime: "fulltime",
  parttime: "parttime",
  contract: "contract",
  remote: "remote",
} as const;

export type JobExperienceLevel =
  (typeof JobExperienceLevel)[keyof typeof JobExperienceLevel];

export const JobExperienceLevel = {
  fresher: "fresher",
  junior: "junior",
  mid: "mid",
  senior: "senior",
} as const;

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  active: "active",
  closed: "closed",
  draft: "draft",
} as const;

export interface Job {
  id: number;
  recruiterId: number;
  recruiterName?: string;
  company: string;
  title: string;
  description: string;
  type: JobType;
  location: string;
  isRemote?: boolean;
  techStack: string[];
  minPay?: number;
  maxPay?: number;
  experienceLevel: JobExperienceLevel;
  requirements?: string[];
  responsibilities?: string[];
  benefits?: string[];
  applicationDeadline?: string;
  openings?: number;
  status: JobStatus;
  applicationsCount?: number;
  createdAt: string;
}

export type JobInputType = (typeof JobInputType)[keyof typeof JobInputType];

export const JobInputType = {
  internship: "internship",
  fulltime: "fulltime",
  parttime: "parttime",
  contract: "contract",
  remote: "remote",
} as const;

export type JobInputExperienceLevel =
  (typeof JobInputExperienceLevel)[keyof typeof JobInputExperienceLevel];

export const JobInputExperienceLevel = {
  fresher: "fresher",
  junior: "junior",
  mid: "mid",
  senior: "senior",
} as const;

export type JobInputStatus =
  (typeof JobInputStatus)[keyof typeof JobInputStatus];

export const JobInputStatus = {
  active: "active",
  closed: "closed",
  draft: "draft",
} as const;

export interface JobInput {
  company: string;
  title: string;
  description: string;
  type: JobInputType;
  location: string;
  isRemote?: boolean;
  techStack: string[];
  minPay?: number;
  maxPay?: number;
  experienceLevel: JobInputExperienceLevel;
  requirements?: string[];
  responsibilities?: string[];
  benefits?: string[];
  applicationDeadline?: string;
  openings?: number;
  status?: JobInputStatus;
}

export interface JobListResponse {
  jobs: Job[];
  total: number;
  page: number;
  totalPages: number;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  reviewing: "reviewing",
  shortlisted: "shortlisted",
  rejected: "rejected",
  offered: "offered",
} as const;

export interface Application {
  id: number;
  jobId: number;
  userId: number;
  applicantName?: string;
  applicantEmail?: string;
  job?: Job;
  coverLetter?: string;
  status: ApplicationStatus;
  appliedAt: string;
  updatedAt?: string;
}

export interface ApplicationInput {
  jobId: number;
  coverLetter?: string;
}

export type ApplicationStatusUpdateStatus =
  (typeof ApplicationStatusUpdateStatus)[keyof typeof ApplicationStatusUpdateStatus];

export const ApplicationStatusUpdateStatus = {
  pending: "pending",
  reviewing: "reviewing",
  shortlisted: "shortlisted",
  rejected: "rejected",
  offered: "offered",
} as const;

export interface ApplicationStatusUpdate {
  status: ApplicationStatusUpdateStatus;
}

export type ResumeGenerateRequestTemplateId =
  (typeof ResumeGenerateRequestTemplateId)[keyof typeof ResumeGenerateRequestTemplateId];

export const ResumeGenerateRequestTemplateId = {
  modern: "modern",
  classic: "classic",
  minimal: "minimal",
  bold: "bold",
} as const;

export interface ResumeGenerateRequest {
  templateId: ResumeGenerateRequestTemplateId;
  includePhoto?: boolean;
}

export interface ResumeResponse {
  html: string;
  templateId: string;
}

export type ListJobsParams = {
  search?: string;
  type?: ListJobsType;
  location?: string;
  techStack?: string;
  minPay?: number;
  maxPay?: number;
  experienceLevel?: ListJobsExperienceLevel;
  page?: number;
  limit?: number;
};

export type ListJobsType = (typeof ListJobsType)[keyof typeof ListJobsType];

export const ListJobsType = {
  internship: "internship",
  fulltime: "fulltime",
  parttime: "parttime",
  contract: "contract",
  remote: "remote",
} as const;

export type ListJobsExperienceLevel =
  (typeof ListJobsExperienceLevel)[keyof typeof ListJobsExperienceLevel];

export const ListJobsExperienceLevel = {
  fresher: "fresher",
  junior: "junior",
  mid: "mid",
  senior: "senior",
} as const;