# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TechPath Jobs & Internships API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job listings with filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [internship, fulltime, parttime, contract, remote]
        - name: location
          in: query
          schema:
            type: string
        - name: techStack
          in: query
          schema:
            type: string
        - name: minPay
          in: query
          schema:
            type: number
        - name: maxPay
          in: query
          schema:
            type: number
        - name: experienceLevel
          in: query
          schema:
            type: string
            enum: [fresher, junior, mid, senior]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Job listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job listing (recruiters only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications for current user
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/status:
    put:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status (recruiter/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        role:
          type: string
          enum: [student, recruiter, admin]
      required: [email, password, name, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, recruiter, admin]
        createdAt:
          type: string
          format: date-time
        profileComplete:
          type: boolean
      required: [id, email, name, role, createdAt, profileComplete]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        recruiterId:
          type: integer
        recruiterName:
          type: string
        company:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [internship, fulltime, parttime, contract, remote]
        location:
          type: string
        isRemote:
          type: boolean
        techStack:
          type: array
          items:
            type: string
        minPay:
          type: number
        maxPay:
          type: number
        experienceLevel:
          type: string
          enum: [fresher, junior, mid, senior]
        requirements:
          type: array
          items:
            type: string
        responsibilities:
          type: array
          items:
            type: string
        benefits:
          type: array
          items:
            type: string
        applicationDeadline:
          type: string
          format: date
        openings:
          type: integer
        status:
          type: string
          enum: [active, closed, draft]
        applicationsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, recruiterId, company, title, description, type, location, techStack, experienceLevel, status, createdAt]
    JobInput:
      type: object
      properties:
        company:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [internship, fulltime, parttime, contract, remote]
        location:
          type: string
        isRemote:
          type: boolean
        techStack:
          type: array
          items:
            type: string
        minPay:
          type: number
        maxPay:
          type: number
        experienceLevel:
          type: string
          enum: [fresher, junior, mid, senior]
        requirements:
          type: array
          items:
            type: string
        responsibilities:
          type: array
          items:
            type: string
        benefits:
          type: array
          items:
            type: string
        applicationDeadline:
          type: string
          format: date
        openings:
          type: integer
        status:
          type: string
          enum: [active, closed, draft]
      required: [company, title, description, type, location, techStack, experienceLevel]
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [jobs, total, page, totalPages]
    Application:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        userId:
          type: integer
        applicantName:
          type: string
        applicantEmail:
          type: string
        job:
          $ref: "#/components/schemas/Job"
        coverLetter:
          type: string
        status:
          type: string
          enum: [pending, reviewing, shortlisted, rejected, offered]
        appliedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, jobId, userId, status, appliedAt]
    ApplicationInput:
      type: object
      properties:
        jobId:
          type: integer
        coverLetter:
          type: string
      required: [jobId]
    ApplicationStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewing, shortlisted, rejected, offered]
      required: [status]
...